refactor: move tokenize property to class config
